new class of pokemons
