+++ Poke_project/src/PokemonComponent.ts
